formatter

+++ contracts/GaiaStore.sol
@@ -20,7 +20,7 @@ contract GaiaStore is Ownable, Pausable {
      * @dev constructor
      */
     constructor(address _nfticket, address[] memory _whitelist) {
-        for(uint i = 0; i < _whitelist.length; i += 1) {
+        for (uint256 i = 0; i < _whitelist.length; i += 1) {
             store.whitelist[_whitelist[i]] = true;
         }
         store.nfticket = IERC721(_nfticket);
@@ -51,15 +51,18 @@ contract GaiaStore is Ownable, Pausable {
         uint256 _ticketPrice,
         uint256[] calldata _tickets
     ) external onlyOwner whenNotPaused {
-        for(uint i = 0; i < _tickets.length; i += 1) {
+        for (uint256 i = 0; i < _tickets.length; i += 1) {
             uint256 _ticketId = _tickets[i];
-            bool gaiaStoreIsOwner = store.nfticket.ownerOf(_ticketId) == address(this);
-            if(!gaiaStoreIsOwner) {
-                string memory errorMsg = string(abi.encodePacked(
-                    "nfticket ",
-                    _ticketId.toString(),
-                    " not owned by gaiastore"
-                ));
+            bool gaiaStoreIsOwner = store.nfticket.ownerOf(_ticketId) ==
+                address(this);
+            if (!gaiaStoreIsOwner) {
+                string memory errorMsg = string(
+                    abi.encodePacked(
+                        "nfticket ",
+                        _ticketId.toString(),
+                        " not owned by gaiastore"
+                    )
+                );
                 revert(errorMsg);
             }
             store.eventTickets[_eventId].push(_ticketId);
@@ -75,12 +78,20 @@ contract GaiaStore is Ownable, Pausable {
     /**
      * @dev payables
      */
-    function buyTicket(uint256 _eventId) external payable onlyWhitelisted whenNotPaused {
-        require(store.events[_eventId].status != EventStatus.Void, "event does not exists");
+    function buyTicket(uint256 _eventId)
+        external
+        payable
+        onlyWhitelisted
+        whenNotPaused
+    {
+        require(
+            store.events[_eventId].status != EventStatus.Void,
+            "event does not exists"
+        );
 
         uint256 _ticketsAmount = store.eventTickets[_eventId].length;
         require(_ticketsAmount > 0, "not enough tickets");
-        
+
         uint256 _ticketPrice = store.events[_eventId].ticketPrice;
         require(msg.value >= _ticketPrice, "not enough money");
 
@@ -91,7 +102,7 @@ contract GaiaStore is Ownable, Pausable {
         store.eventTickets[_eventId].pop();
     }
 
-    function getEvent(uint256 _eventId) external view returns(Event memory) {
+    function getEvent(uint256 _eventId) external view returns (Event memory) {
         return store.events[_eventId];
     }
 }
